implement signup page, modulated main page rendering
