+++ client/public/style.css
@@ -1,10 +1,11 @@
 .navBar, .sideBar, .mainPage, .sideBarItem, .footer, .tournaments {
-  border: 3px blue solid;
+  border: 1px lightgray dotted;
 }
 
-.navBar, .sideBar {
-  /* position: fixed; */
-  float: left;
+#navigation {
+  top: 0;
+  left: 0;
+  right: 0;
 }
 
 .carouselItem {
@@ -14,9 +15,12 @@
   z-index: 1;
 }
 
+.static {
+  position: fixed;
+}
 
 .sideBarItem {
-  height: 150px;
+  height: auto;
 }
 
 .carouselImg {
